requirements.txt
absl-py==0.7.1
astor==0.7.1
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2019.3.9
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
html5lib==0.9999999
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1
mock==3.0.5
numpy==1.16.4
opencv-python==4.1.0.25
protobuf==3.5.2
pygame==1.9.6
PyYAML==5.1
rpg-world==1.0
scipy==1.2.1
six==1.12.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
webencodings==0.5.1
Werkzeug==0.15.2
wincertstore==0.2
Pillow